+++ src/Everzet/Behat/Collector/StatisticsCollector.php
@@ -23,7 +23,12 @@ use Everzet\Behat\Tester\StepTester;
 class StatisticsCollector implements CollectorInterface
 {
     protected $paused               = false;
+    protected $suiteStartTime;
+    protected $suiteFinishTime;
+    protected $scenarioStartTime;
+    protected $scenarioFinishTime;
 
+    protected $isPassed             = true;
     protected $statuses             = array(
         StepTester::PASSED      => 'passed'
       , StepTester::SKIPPED     => 'skipped'
@@ -55,6 +60,52 @@ class StatisticsCollector implements CollectorInterface
             array_values($this->statuses)
           , array_fill(0, count($this->statuses), 0)
         );
+    }
+
+    /**
+     * Start suite timer. 
+     */
+    public function startTimer()
+    {
+        $this->suiteStartTime = microtime(true);
+    }
+
+    /**
+     * Stop suite timer. 
+     */
+    public function finishTimer()
+    {
+        $this->suiteFinishTime = microtime(true);
+    }
+
+    /**
+     * Return suite total execution time. 
+     * 
+     * @return  integer miliseconds
+     */
+    public function getTotalTime()
+    {
+        return $this->suiteFinishTime - $this->suiteStartTime;
+    }
+
+    /**
+     * Return last scenario execution time. 
+     * 
+     * @return  integer miliseconds
+     */
+    public function getLastScenarioTime()
+    {
+        return $this->scenarioFinishTime - $this->scenarioStartTime;
+    }
+
+    /**
+     * Return true if suite passed. 
+     * 
+     * @return  boolean
+     */
+    public function isPassed()
+    {
+        return $this->isPassed;
     }
 
     /**
@@ -132,9 +183,11 @@ class StatisticsCollector implements CollectorInterface
      */
     public function registerListeners(EventDispatcher $dispatcher)
     {
-        $dispatcher->connect('scenario.run.after',      array($this, 'collectScenarioStats'),   5);
-        $dispatcher->connect('outline.sub.run.after',   array($this, 'collectScenarioStats'),   5);
-        $dispatcher->connect('step.run.after',          array($this, 'collectStepStats'),       5);
+        $dispatcher->connect('scenario.run.before',     array($this, 'startScenarioTimer'),     10);
+        $dispatcher->connect('outline.sub.run.before',  array($this, 'startScenarioTimer'),     10);
+        $dispatcher->connect('scenario.run.after',      array($this, 'collectScenarioStats'),   10);
+        $dispatcher->connect('outline.sub.run.after',   array($this, 'collectScenarioStats'),   10);
+        $dispatcher->connect('step.run.after',          array($this, 'collectStepStats'),       10);
     }
 
     /**
@@ -154,6 +207,16 @@ class StatisticsCollector implements CollectorInterface
     }
 
     /**
+     * Listens to `scenario.run.before` & `outline.sub.run.before` events & start scenario timers. 
+     * 
+     * @param   Event   $event  event
+     */
+    public function startScenarioTimer(Event $event)
+    {
+        $this->scenarioStartTime = microtime(true);
+    }
+
+    /**
      * Listens to `scenario.run.after` & `outline.sub.run.after` events & collect stats.
      *
      * @param   Event   $event  event
@@ -162,8 +225,13 @@ class StatisticsCollector implements CollectorInterface
     {
         if (!$this->paused) {
             ++$this->scenariosCount;
-            ++$this->scenariosStatuses[$this->statuses[$event['result']]];
+            ++$this->scenariosStatuses[$this->statuses[$event->getParameter('result')]];
+            if (0 !== $event->getParameter('result')) {
+                $this->isPassed = false;
+            }
         }
+
+        $this->scenarioFinishTime = microtime(true);
     }
 
     /**
@@ -175,21 +243,24 @@ class StatisticsCollector implements CollectorInterface
     {
         if (!$this->paused) {
             ++$this->stepsCount;
-            ++$this->stepsStatuses[$this->statuses[$event['result']]];
+            ++$this->stepsStatuses[$this->statuses[$event->getParameter('result')]];
+            if (0 !== $event->getParameter('result')) {
+                $this->isPassed = false;
+            }
 
-            if (StepTester::UNDEFINED === $event['result']) {
-                foreach ($event['snippet'] as $key => $snippet) {
+            if (StepTester::UNDEFINED === $event->getParameter('result')) {
+                foreach ($event->getParameter('snippet') as $key => $snippet) {
                     if (!isset($this->definitionsSnippets[$key])) {
                         $this->definitionsSnippets[$key] = $snippet;
                     }
                 }
             }
 
-            if (StepTester::FAILED === $event['result']) {
+            if (StepTester::FAILED === $event->getParameter('result')) {
                 $this->failedStepsEvents[] = $event;
             }
 
-            if (StepTester::PENDING === $event['result']) {
+            if (StepTester::PENDING === $event->getParameter('result')) {
                 $this->pendingStepsEvents[] = $event;
             }
         }