+++ src/Everzet/Behat/Tester/OutlineTester.php
@@ -33,7 +33,7 @@ class OutlineTester implements NodeVisitorInterface
     public function __construct(Container $container)
     {
         $this->container    = $container;
-        $this->dispatcher   = $container->getBehat_EventDispatcherService();
+        $this->dispatcher   = $container->get('behat.event_dispatcher');
     }
 
     /**
@@ -51,17 +51,18 @@ class OutlineTester implements NodeVisitorInterface
 
         // Run subscenarios of outline based on examples
         foreach ($outline->getExamples()->getTable()->getHash() as $iteration => $tokens) {
-            $this->dispatcher->notify(new Event($outline, 'outline.sub.run.before', array(
-                'iteration' => $iteration
-            )));
-
-            $environment    = $this->container->getBehat_EnvironmentService();
+            $environment    = $this->container->get('behat.environment');
             $itResult       = 0;
             $skip           = false;
 
+            $this->dispatcher->notify(new Event($outline, 'outline.sub.run.before', array(
+                'iteration'     => $iteration
+              , 'environment'   => $environment
+            )));
+
             // Visit & test background if has one
             if ($outline->getFeature()->hasBackground()) {
-                $tester = $this->container->getBehat_BackgroundTesterService();
+                $tester = $this->container->get('behat.background_tester');
                 $tester->setEnvironment($environment);
 
                 $bgResult = $outline->getFeature()->getBackground()->accept($tester);
@@ -74,7 +75,7 @@ class OutlineTester implements NodeVisitorInterface
 
             // Visit & test steps
             foreach ($outline->getSteps() as $step) {
-                $tester = $this->container->getBehat_StepTesterService();
+                $tester = $this->container->get('behat.step_tester');
                 $tester->setEnvironment($environment);
                 $tester->setTokens($tokens);
                 $tester->skip($skip);
@@ -88,9 +89,10 @@ class OutlineTester implements NodeVisitorInterface
             }
 
             $this->dispatcher->notify(new Event($outline, 'outline.sub.run.after', array(
-                'iteration' => $iteration
-              , 'result'    => $itResult
-              , 'skipped'   => $skip
+                'iteration'     => $iteration
+              , 'result'        => $itResult
+              , 'skipped'       => $skip
+              , 'environment'   => $environment
             )));
 
             $result = max($result, $itResult);